Incuding original error w/ custom for invalid cert
Between version 6 and 8 there is a difference in how the invalid cert is
raised. In version 8 it's now an OpenSSL error that needs handled
differently. The commit prior to this checks for the existance of the
new error message. This commit focuses on including the original error
with our custom one for some further inspection upstream. It also
increases clarity in what is being tested and what has changed in the
PR.